Add support for regress tests
Adding support for running `pg_regress` when creating an extension by
providing a list of test files to run for the extension.

Also added custom commands `test_verbose` to show the full output of the
test (including the `regression.diff`) and a target
`<extension>_update_results` to update the result files.